refactor: rename `getNodeType` to `getAnimatedType` (and move it to packages/animated)
