+++ packages/animated/src/index.ts
@@ -4,5 +4,6 @@ export * from './AnimatedString'
 export * from './AnimatedArray'
 export * from './AnimatedObject'
 export * from './AnimatedProps'
+export * from './getAnimatedType'
 export * from './createHost'
 export * from './types'
